refactor(db-types): extract constants to fix max-lines lint
