+++ packages/db-types/src/index.ts
@@ -212,11 +212,5 @@ export type NudgeLogRow = InferSelectModel<typeof nudgeLog>;
 export type NudgeLogInsert = InferInsertModel<typeof nudgeLog>;
 
 // Constants
-export const ENTITY_TYPES = ['company', 'person', 'place', 'brand', 'organisation'] as const;
-export type EntityType = (typeof ENTITY_TYPES)[number];
-
-export const WISH_LIST_PRIORITIES = ['Needing', 'Soon', 'One Day', 'Dreaming'] as const;
-export type WishListPriority = (typeof WISH_LIST_PRIORITIES)[number];
-
-export const MEDIA_TYPES = ['movie', 'tv_show'] as const;
-export type MediaType = (typeof MEDIA_TYPES)[number];
+export { ENTITY_TYPES, WISH_LIST_PRIORITIES, MEDIA_TYPES } from './constants.js';
+export type { EntityType, WishListPriority, MediaType } from './constants.js';
